% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDensity.R
\name{calcDensity}
\alias{calcDensity}
\title{Calculate density in a given region}
\usage{
calcDensity(pos_info, region_info, slide_window = 1e+05)
}
\arguments{
\item{pos_info}{tab-separated file (or a data frame) containing position information of items to calculate density, with first column being "ID", second being "Pos". File should NOT have header.}

\item{region_info}{tab-separated file (or a data frame) containing information about the full-length region to calculate density in, with first column being "ID", second being "Start", third being "End". File should NOT have header.}

\item{slide_window}{length of each slide window, default 100kb}
}
\description{
Calculate density in a given region with a slide window
}
\author{
Yujie Liu
}
