% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDensityGFF.R
\name{calcDensityGFF}
\alias{calcDensityGFF}
\title{Calculate gene density on chromosome}
\usage{
calcDensityGFF(gff_file, karyotype_info, feature = "gene", window = 1e+05)
}
\arguments{
\item{gff_file}{GFF file}

\item{karyotype_info}{tab-separated file (or a data frame) containing karyotype information, with first column being "Chr", second being "Start", third being "End". File should NOT have header.}

\item{feature}{interested feature to calculate, should be in GFF file}

\item{window}{length of each slide window, defaut 100kb}
}
\description{
Calculate density of genes, CDS, mRNAs, tRNAs, etc on chromosome from a given GFF file with a slide window
}
\author{
Yujie Liu
}
