% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandDups.R
\name{expandDups}
\alias{expandDups}
\title{Expand one entry into multiple lines}
\usage{
expandDups(x, id = 1, sep = ",")
}
\arguments{
\item{x}{a data frame or tibble}

\item{id}{column index of variable needs to expand, which has separators defined as \code{sep}. Default the first column.}

\item{sep}{separator}
}
\description{
Expand one entry, which is separated into different part (which may have different meanings) by a common separator, into multiple lines. Note: this function may destroy your data in some way, since its simply "copy" the rest columns to each entry, so you should be well aware of the meaning of your data.
}
\examples{
x <-
  data.frame(
    x1 = c("A,A,A", "B,B", "C"),
    x2 = c(1, 2, 3),
    x3 = c("100,100,200", "200,200", "300")
  )
expandDups(x, id = 1)
}
\author{
Yujie Liu
}
