% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPos.R
\name{extractPos}
\alias{extractPos}
\title{Extract All Elements from GFF3 File}
\usage{
extractPos(gff_file, chr_len_file, bed_file)
}
\arguments{
\item{gff_file}{GFF3 file}

\item{chr_len_file}{file containing chromosome length}

\item{bed_file}{BED file for output}
}
\description{
Extract all elements (promoter, UTR, CDS, intron) of each gene from GFF3 file into BED format
}
\author{
Yujie Liu
}
