% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGC.R
\name{fastGC}
\alias{fastGC}
\title{Compute GC content across all reads in fastq files}
\usage{
fastGC(file, start = 1L, end = 1000L, verbose = FALSE)
}
\arguments{
\item{file}{fastq file, not in gzip}

\item{start}{visualization start location, default the first read}

\item{end}{visualization end location, default the 1000th read}

\item{verbose}{whether to display the result on the screen}
}
\description{
Compute GC content across all reads in fastq files, which mainly derived from high-throughput next-generation sequencing.
}
\author{
Yujie Liu
}
