% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastIN.R
\name{fastIN}
\alias{fastIN}
\title{Read fasta/fastq files}
\usage{
fastIN(file, format)
}
\arguments{
\item{file}{input file name (path)}

\item{format}{input file format, must be one of "fasta" or "fastq".}
}
\description{
Read fasta or fastq file into a data frame. For FASTA files, it will contain sequence names (without ">") in the first column, and sequence content in the second. For FASTQ files, it will contain sequence information (without "@") in the first column, sequence content in the second, additional information (without "+") in the third, and quality information in the fourth.
}
\author{
Yujie Liu
}
