% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastQCer.R
\name{fastQCer}
\alias{fastQCer}
\title{Quality Check for fastq files}
\usage{
fastQCer(file, start = 1L, end = 1000L)
}
\arguments{
\item{file}{fastq file, not in gzip}

\item{start}{QC start location, default the first read}

\item{end}{QC end location, default the 1000th read}
}
\description{
Quality Check for fastq files, which mainly derived from high-throughput next-generation sequencing. It is not just a collection of fastQual, fastGC, fastSC, fastSQ, etc., some new features are also added.
}
\author{
Yujie Liu
}
