% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastQual.R
\name{fastQual}
\alias{fastQual}
\title{Visualize quality information in fastq files}
\usage{
fastQual(file, mode = "text", start = 1L, end = 1000L, verbose = FALSE)
}
\arguments{
\item{file}{fastq file, not in gzip}

\item{mode}{visualization mode, now support "text" and "plot"}

\item{start}{visualization start location, default the first read}

\item{end}{visualization end location, default the 1000th read}

\item{verbose}{whether to display the result on the screen}
}
\description{
Visualize quality information in fastq files, which mainly derived from high-throughput next-generation sequencing. There are two modes: in "text" mode, it will generate text-based bar plots of relative possibility of error (the lower the better); in "plot" mode, it will return a plot displaying per base sequence quality (the higher the better).
}
\author{
Yujie Liu
}
