% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSC.R
\name{fastSC}
\alias{fastSC}
\title{Compute sequence content across all bases in fastq files}
\usage{
fastSC(file, start = 1L, end = 1000L, verbose = FALSE)
}
\arguments{
\item{file}{fastq file, not in gzip}

\item{start}{visualization start location, default the first read}

\item{end}{visualization end location, default the 1000th read}

\item{verbose}{whether to display the result on the screen}
}
\description{
Compute sequence content across all bases in fastq files, which mainly derived from high-throughput next-generation sequencing.
}
\author{
Yujie Liu
}
