% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSQ.R
\name{fastSQ}
\alias{fastSQ}
\title{Mean sequence quality scores across all reads in fastq files}
\usage{
fastSQ(file, start = 1L, end = 1000L, verbose = FALSE)
}
\arguments{
\item{file}{fastq file, not in gzip}

\item{start}{visualization start location, default the first read}

\item{end}{visualization end location, default the 1000th read}

\item{verbose}{whether to display the result on the screen}
}
\description{
Compute mean sequence quality scores all reads in fastq files, which mainly derived from high-throughput next-generation sequencing.
}
\author{
Yujie Liu
}
