% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fopen.R
\name{fopen}
\alias{fopen}
\title{Open file with system default application}
\usage{
fopen(file)
}
\arguments{
\item{file}{file to open}
}
\description{
Open file with system default application.
}
\author{
Yujie Liu
}
