% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genometracks.R
\name{genometracks}
\alias{genometracks}
\title{Plot Genome Tracks for NGS Data}
\usage{
genometracks(
  bed_file,
  track_conf,
  gene_model,
  ann_region = NULL,
  ann_color = "goldenrod1",
  out_format = "png",
  out_dir = "out"
)
}
\arguments{
\item{bed_file}{BED file containing locus to plot. For each line, TAB-delimited information should containing: <chr>, <start>, <end>, <locus_name>.}

\item{track_conf}{configuration file containing track info. For each line, TAB-delimited information should containing: <track_name>, <path_to_track_file>, <y_min>, <y_max>, <color>, <group>, without header.}

\item{gene_model}{BED file containing all elements of gene and TE model, generated by \verb{[tinyfuncr::extractPos]}}

\item{ann_region}{BED file of annotation regions. If not provided, no annotation will be added. For each line, TAB-delimited information should containing: <chr>, <start>, <end>, <annotation_name>. <annotation_name> could be empty but the last TAB should be kept.}

\item{ann_color}{color of annotated regions}

\item{out_format}{format of output figures, one of "png" and "pdf"}

\item{out_dir}{path to output directory, create it if not exist}
}
\description{
Plot genome tracks for NGS data, resembling GBrowse/JBrowse style. Supporting coverage data in bigWig format from RNA-seq, sRNA-seq, BS-seq, ChIP-seq, RIP-seq, GRO-seq, NET-seq, DNase-seq, MNase-seq, ATAC-seq, SHAPE-seq etc. Annotation should be provided in GTF format. Some codes refers to https://github.com/PoisonAlien/trackplot and \code{genomemodel} package with some major modifications. Requires \code{bedtools} and \code{bwtool} on Linux.
}
\author{
Yujie Liu
}
