% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getigvtable.R
\name{getigvtable}
\alias{getigvtable}
\title{Get IGV online reference genomes table}
\usage{
getigvtable()
}
\description{
Get IGV pre-defined reference genomes and annotation. It will return a table containing current list of pre-defined reference genomes' names and IDs. Note: You should only use IDs available in this table in the function \code{makeigvref}.
}
\examples{
#getigvtable()
}
\author{
Yujie Liu
}
