% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggepitracks.R
\name{ggepitracks}
\alias{ggepitracks}
\title{Plot Epigenetics Tracks for NGS Data using ggplot2}
\usage{
ggepitracks(
  locus_file,
  track_conf,
  bsseq_conf = NULL,
  gene_model,
  ann_region = NULL,
  ann_color = "goldenrod1",
  out_dir = "out"
)
}
\arguments{
\item{locus_file}{BED file containing locus to plot. For each line, TAB-delimited information should contain: <chr>, <start>, <end>, <locus_name>, without header.}

\item{track_conf}{configuration file containing track info. For each line, TAB-delimited information should contain: <track_name>, <track_type>, <track_file>, <y_min>, <y_max>, <color>, <group>, without header. NOTE: If you write "BS-seq" in "<track_type>", provide mC type (one of CG, CHG, CHH, All) to plot in "<path_to_track_file>" and provide detailed info in \code{bsseq_conf}, "<color>" is currently ignored in multi-mC mode of BS-seq.}

\item{bsseq_conf}{configuration file of BS-seq track. If no BS-seq track plotted, keep \code{NULL}. For each line, TAB-delimited information should contain: <track_name>, <CG_file>, <CHG_file>, <CHH_file>, <C_context_file>, without header. "<track_name>" here should be the same as in \code{track_conf}. If you choose only one mC type to plot, you can keep the column of other file as "NA", but provide all the file may be used has no side-effect. "<C_context_file>" can be "NA" to avoid drawing C context track.}

\item{gene_model}{BED file containing all elements of gene and/or TE model, generated by \verb{[tinyfuncr::extractPos]}}

\item{ann_region}{BED file of annotation regions. If not provided, no annotation will be added. For each line, TAB-delimited information should contain: <chr>, <start>, <end>, <annotation_name>, without header. <annotation_name> could be empty but the last TAB should be kept.}

\item{ann_color}{color of annotated regions}

\item{out_dir}{path to output directory, create it if not exist}
}
\description{
Plot genome tracks for NGS data using ggplot2, resembling GBrowse/JBrowse style. Supporting coverage data in bigWig format from RNA-seq, sRNA-seq, BS-seq, ChIP-seq, RIP-seq, GRO-seq, NET-seq, DNase-seq, MNase-seq, ATAC-seq, SHAPE-seq etc. Especially optimized for BS-seq data in single-base resolution. Requires \code{bedtools} on Linux.
}
\author{
Yujie Liu
}
