% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfanplot.R
\name{ggfanplot}
\alias{ggfanplot}
\title{Add "fanplot" to chromosome}
\usage{
ggfanplot(
  p,
  plot_type = c("text", "shape", "textshape"),
  plot_direction = c("vertical", "horizontal"),
  fan_start = 0.8,
  fan_end = 3,
  tip_length = 0.2,
  ann_data = NULL,
  htext_direction = c("right", "left"),
  replusive = 1e+06,
  text_size = 1,
  text_color = "black",
  shape_size = 1,
  shape_color = "skyblue",
  shape_repel = 0.25,
  ...
)
}
\arguments{
\item{p}{a ggplot object}

\item{plot_type}{type of fanplot annotation, one of "text", "shape", and "textshape"}

\item{plot_direction}{direction of chromosome, one of "vertical" and "horizontal"}

\item{fan_start}{"fan" start position}

\item{fan_end}{"fan" end position}

\item{tip_length}{"fan" tip length, both sides are the same}

\item{ann_data}{a data frame containing annotation information, colnames should be set as following:
\itemize{
\item For \strong{text} type (in \code{plot_type}), first column is "Chr", second is "Pos", third is "Text", fouth is "TextSize", fifth is "TextColor", where fouth and fifth column is optional. When "TextSize" and "TextColor" are not given to each annotation, use global options from \code{text_size} and \code{text_color}.
\item For \strong{shape} type, first column is "Chr", second is "Pos", third is "Shape", fouth is "ShapeSize", fifth is "ShapeColor", where fouth and fifth column is optional. When "ShapeSize" and "ShapeColor" are not given to each shape, use global options from \code{shape_size} and \code{shape_color}.
\item For \strong{textshape} type, first column is "Chr", second is "Pos", third is "Text", fouth is "Shape", fifth is "TextSize", sixth is "TextColor", seventh is "ShapeSize", eighth is "ShapeColor", where 5-8 columns are optional.
}}

\item{htext_direction}{from which side to read the text when chromosomes lying down (i.e. \code{plot_direction = "horizontal"}), one of "right" and "left"}

\item{replusive}{replusive of texts or shapes, set according to chr length}

\item{text_size}{global text size}

\item{text_color}{global text color}

\item{shape_size}{global shape size}

\item{shape_color}{global shape color}

\item{shape_repel}{shape repulsive from fan end, absolute length against x/y axis}

\item{...}{currently ignored}
}
\description{
Add "fanplot" to chromosome. "fanplot" means repel annotation lines linking between original positions and annotations, named after the fan made by paper due to their similarity in appearance.
}
\author{
Yujie Liu
}
