% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeapCal.R
\name{LeapCal}
\alias{LeapCal}
\title{Leap-year Calendar}
\usage{
LeapCal(year, month)
}
\arguments{
\item{year}{year, such as 2016.}

\item{month}{month, such as 08 or 8.}
}
\description{
A leap-year calendar written in C++ and implemented in R with Rcpp. Just a trial in using Rcpp.
}
\author{
Yujie Liu
}
