% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeigvref.R
\name{makeigvref}
\alias{makeigvref}
\title{Make IGV reference genome information}
\usage{
makeigvref(refid)
}
\arguments{
\item{refid}{reference genome id}
}
\description{
Make reference genome information used in IGViewer.
}
\examples{
makeigvref("hg38")
}
\author{
Yujie Liu
}
