% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDups.R
\name{mergeDups}
\alias{mergeDups}
\title{Merge duplicated lines into one entry}
\usage{
mergeDups(x, id, mer = -id)
}
\arguments{
\item{x}{a data frame or tibble}

\item{id}{column index of identical variable, which is used to detect duplication.}

\item{mer}{column index of variables to merge, default the rest. These columns should only contain numeric values, because merging means simply adding these numbers together.}
}
\description{
Merge duplicated lines into one entry. Useful especially in manipulating files about SNP annotation result generated by snpEff and corresponding downstream analysis, where duplicated entries were generated due to multiple transcripts in one gene or multiple samples in one file. Note: this function may be quite slow when input data is large.
}
\examples{
x <-
  data.frame(
    x1 = c("A", "A", "B", "A", "C", "B"),
    x2 = c(1, 1, 2, 1, 3, 2),
    x3 = c(100, 100, 200, 100, 300, 200),
    x4 = c(100, 200, 300, 400, 500, 600)
  )

merged1 <- mergeDups(x, id = 1)
merged2 <- mergeDups(x, id = 1:3)
}
\author{
Yujie Liu
}
