% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWrapper.R
\name{RWrapper}
\alias{RWrapper}
\alias{read_tcsv}
\alias{write_tcsv}
\title{Wrapper for reading/writing table/csv files}
\usage{
read_tcsv(
  file,
  type = "tsv",
  header = TRUE,
  quote = "",
  sep = NULL,
  row.names = NULL,
  col.names = NULL,
  comment.char = "",
  stringsAsFactors = FALSE,
  fileEncoding = "UTF-8"
)

write_tcsv(
  x,
  file = "tmp",
  type = "tsv",
  append = FALSE,
  quote = FALSE,
  sep = NULL,
  col.names = TRUE,
  row.names = FALSE,
  fileEncoding = "UTF-8"
)
}
\arguments{
\item{file}{file's full name to read in or write to}

\item{type}{file type, either "tsv" or "csv"}

\item{x}{typically a data frame or tibble}

\item{...}{see more at \code{\link[utils:read.table]{utils::read.table()}} and \code{\link[utils:write.table]{utils::write.table()}}}
}
\description{
Wrapper for reading/writing table/csv files with sensible defaults.
}
\author{
Yujie Liu
}
