% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam2bed.R
\name{sam2bed}
\alias{sam2bed}
\title{Convert SAM/BAM file to BED}
\usage{
sam2bed(input, output, gr = FALSE)
}
\arguments{
\item{input}{input SAM/BAM file}

\item{output}{output BED file's name, without suffix}

\item{gr}{whether to generate a GRange object, default FALSE}
}
\description{
Convert SAM (Sequence Alignment/Mapping) or BAM (binary form of SAM) file to BED (Browser Extensible Data) file. SAM files can only contain single-end-reads now.
}
\examples{
samfile <- system.file(
              "extdata",
              "THA2.sam",
              package = "tinyfuncr",
              mustWork = TRUE
           )
bamfile <- system.file(
              "extdata",
              "THA2_sorted.bam",
              package = "tinyfuncr",
              mustWork = TRUE
           )

sam2bed(input = samfile, output = "THA2")
gr <- sam2bed(input = bamfile,
              output = "THA2_sorted",
              gr = TRUE)
}
\author{
Yujie Liu
}
