% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toUTF8.R
\name{toUTF8}
\alias{toUTF8}
\title{Convert file(s) with different encoding(s) to UTF-8}
\usage{
toUTF8(file, return = FALSE)
}
\arguments{
\item{file}{input file(s)}

\item{return}{whether to generate a character vector containing all the files' content}
}
\description{
Convert files with different encoding(s) to UTF-8 and then processing them with various R functions.
}
\author{
Yujie Liu
}
