% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm2md.R
\name{bm2md}
\alias{bm2md}
\title{Convert html format bookmarks to markdown}
\usage{
bm2md(
  file = dir(pattern = ".html$")[1],
  out = "bookmarks",
  safari = FALSE,
  add.rules = NULL
)
}
\arguments{
\item{file}{bookmarks file in html format. Default: first html file in the current directory}

\item{out}{output markdown file's name, without suffix}

\item{safari}{whether this file is exported from MacOS Safari, default FALSE}

\item{add.rules}{additional rules for some strange items. It should be a tibble or data frame, with first column being \code{pattern}, second being \code{replacement} (used in function \code{stringr::str_replace_all()}). All items should be written in string form of regular expression.}
}
\description{
Convert html format bookmarks to markdown. Useful especially in publishing it on GitHub. Only bookmarks files exported from Microsoft Edge, Google Chrome and Safari are supported now.
}
\examples{

### Example additional rules

add.rules <- data.frame(
  pattern = c('.+HREF="(.+)" (LAST|ADD).+>(.+)',
              "test"),
  replacement = c("* [\\\\3](\\\\1)",
                  "TEST")
)

}
\author{
Yujie Liu
}
