% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDups.R
\name{mergeDups}
\alias{mergeDups}
\title{Merge duplicated lines into one entry}
\usage{
mergeDups(x, id = 1, mer = 2:ncol(x))
}
\arguments{
\item{x}{a data frame or tibble}

\item{id}{column index of identical variable, which is used to detect duplication, default 1}

\item{mer}{column index of variables to merge, default the rest (2:ncol(x)). These columns should only contain numeric values, because merging means simply adding these numbers together.}
}
\description{
Merge duplicated lines into one entry. Useful especially in manipulating files about SNP annotation result generated by snpEff, where duplicated entries were generated due to multiple transcripts in one gene.
}
\examples{
x <-
  data.frame(
    a = c(1, 2, 3, 2, 3, 2, 4, 5),
    b = c(1, 2, 3, 2, 3, 2, 4, 5),
    c = c(1, 2, 3, 2, 3, 2, 4, 5),
    d = c(1, 2, 3, 2, 3, 2, 4, 5)
  )

x1 <- mergeDups(x, id = 1, mer = 2)
x2 <- mergeDups(x, id = 1, mer = 2:3)
x3 <- mergeDups(x, id = 1, mer = c(2,4))
}
\author{
Yujie Liu
}
