% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsave.R
\name{plotsave}
\alias{plotsave}
\title{Save plot to device with various formats}
\usage{
plotsave(
  plot = NULL,
  type = "pdf",
  filename = NULL,
  height = 6,
  width = 6,
  open = FALSE,
  ...
)
}
\arguments{
\item{plot}{a given plot object or current plot (default).}

\item{type}{file type, in "pdf", "eps", "svg", "ppt(x)", "doc(x)", "png", "jp(e)g", "tif", "html". Note: "html" type work for ggplot2 object only.}

\item{filename}{file name, without suffix. Default: combines "Rplot" and current time together}

\item{height}{graph height}

\item{width}{graph width}

\item{open}{whether to open graph file after saving, default FALSE}

\item{...}{other arguments used in package "export"}
}
\description{
Save the current graph or a specific plot item to device.
}
\examples{
df <- data.frame(x = 1:5, y = 6:10)
plot(df)
g <- ggplot(df) + geom_point(aes(x, y))
plotsave(open = TRUE)
plotsave(g, "pptx", "myggplot")
}
\author{
Yujie Liu
}
