% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{spirograph}
\alias{spirograph}
\title{Draw Spirograph}
\usage{
spirograph(R = 25, r = 12, p = 5, col = "skyblue", width = 0.2)
}
\arguments{
\item{R}{outer circle radius, should only contain one value}

\item{r}{numeric vector of inner circle radius}

\item{p}{numeric vector of distance between pen-point and center of inner circle}

\item{col}{character vector of color of each line}

\item{width}{numeric vector of width of each line. Typically, \code{r}, \code{p}, \code{col}, and \code{width} should be the same length}
}
\description{
Draw Spirograph
}
\examples{
spirograph(31, seq(11, 20, 2), seq(6, 15, 2))

}
\author{
Yujie Liu
}
